@KevinWinkler-CN-Consult fixed issue where the shape drops one more space after the pause button is pressed and added a cover that stops the board from being seen while the game is paused.

+++ Tetris/Scripts/UI.gd
@@ -18,27 +18,36 @@ var pressed := 0
 var score := 0
 var ten_lines := 0
 var previous_ten_lines := 0
-
+var remainingDropTime := 0
+signal changePauseState
 
 func _ready():
 	# displays the highscore when the game starts and hides the game over label
 	set_highscore()
 	over.hide()
 
+func _process(delta):
+	if Input.is_action_just_pressed("pause"):
+		_on_Pause_pressed()
 
 func _on_Pause_pressed() -> void:
-	# pauses the game when pressed for the first time, continues it when pressed for the second time
-	pause.disabled = true
-	active_shape.active = false
-	pressed += 1
-	pause.texture_normal = UNPAUSE_RED_IMG
-	if pressed == 2:
-		pause.texture_normal = PAUSE_RED_IMG
-		active_shape.active = true
-		active_shape.drop()
-		pressed = 0
-	# starts a two second timer between pause button clicks to prevent errors and spam clicking
-	pause_timer.start()
+	if !pause.disabled:
+		# pauses the game when pressed for the first time, continues it when pressed for the second time
+		pause.disabled = true
+		remainingDropTime = active_shape.drop_timer.wait_time
+		active_shape.active = false
+		emit_signal("changePauseState",true)
+		pressed += 1
+		pause.texture_normal = UNPAUSE_RED_IMG
+		if pressed == 2:
+			pause.texture_normal = PAUSE_RED_IMG
+			active_shape.active = true
+			emit_signal("changePauseState",false)
+			active_shape.drop_timer.wait_time = remainingDropTime
+			active_shape.drop()
+			pressed = 0
+		# starts a two second timer between pause button clicks to prevent errors and spam clicking
+		pause_timer.start()
 
 
 func _on_PauseTimer_timeout() -> void:
